backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.1
python-dotenv==1.0.1
langchain==0.3.9
langchain-chroma==0.1.4
chromadb==0.5.20
sentence-transformers==3.3.1
langchain-community>=0.3.0
langchain-huggingface>=0.3.0
groq>=1.0.0
